+++ src/controllers/users-controller.js
@@ -1,5 +1,6 @@
 const bcrypt = require("bcrypt");
 const User = require("../models/user-model.js");
+const jwt = require("jsonwebtoken");
 
 const createUser = async (req, res) => {
     var regex = new RegExp(/^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/g, "g")
@@ -19,11 +20,19 @@ const createUser = async (req, res) => {
 }
 
 const getUser = async (req, res) => {
+    if (req.user.role !== "admin" && req.user.id !== req.params.id) {
+        return res.status(403).json({ message: "Non autorisé à consulter ce compte" });
+    }
+
     const user = await User.findById(req.params.id);
     return res.status(200).json(user);
 }
 
 const updateUser = async (req, res) => {
+    if (req.user.role !== "admin" && req.user.id !== req.params.id) {
+        return res.status(403).json({ message: "Non autorisé à consulter ce compte" });
+    }
+    
     const id = req.params.id;
     const user = await User.findById(id);
     const newEmailUser = await User.findOne({ email: req.body.email })
@@ -44,15 +53,58 @@ const updateUser = async (req, res) => {
 }
 
 const deleteUser = async (req, res) => {
+    if (req.user.role !== "admin" && req.user.id !== req.params.id) {
+        return res.status(403).json({ message: "Non autorisé à consulter ce compte" });
+    }
+    
     const id = req.params.id;
     const user = await User.findByIdAndDelete(id);
 
     res.status(200).json({ message: `L'utilisateur ${id} est désintégré !`, user });
 }
 
+const loginUser = async (req, res) => {
+    const { email, password } = req.body;
+    const user = await User.findOne({ email });
+    User.findOne({email})
+        .then(
+            (async user => {
+                if (!user) {
+                    return res.status(404).json({ error: "Utilisateur non trouvé" });
+                }
+                
+                const match = await bcrypt.compare(password, user.password);
+                if (!match) {
+                    return res.status(403).json({ error: "Mot de passe incorrect" });
+                }
+
+                const token = jwt.sign(
+                    { 
+                        id: user._id, 
+                        role: user.role, 
+                        email: user.email 
+                    },
+                    process.env.JWT_SECRET || "superSecretKey", // a mettre dans un .env
+                    { expiresIn: "1h" }
+                );
+
+                return res.status(200).json({ 
+                    message: `Salut ${user.username}, tu es connecté !`,
+                    token
+                });
+            })
+        )
+        .catch(
+            (error => {
+                res.status(500).json(error);
+            })
+        )
+}
+
 module.exports = {
     createUser,
     getUser,
     updateUser,
-    deleteUser
+    deleteUser,
+    loginUser
 }
